# app/web/swagger-ui/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Parkimeter Affiliates Reporting API",
        "description": "This API is meant to provide developers a rapid way to discover the affiliate tracking data",
        "version": "1.0.0"
    },
    "host": "192.168.10.12",
    "basePath": "/app_dev.php",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/cancellation-tracks/": {
            "get": {
                "tags": [
                    "Cancellation Tracks"
                ],
                "summary": "Returns a paginated list of cancellation tracks.",
                "operationId": "getManyCancellationTrack",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "per-page",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "affiliate-id",
                        "in": "query",
                        "description": "Filter by affiliate-id",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "from-date",
                        "in": "query",
                        "description": "Filter by a date of being",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "to-date",
                        "in": "query",
                        "description": "Filter by a date of end",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/SwaggerGetManyCancellationTrackSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "SwaggerGetCancellationTrackSuccessResponse": {
            "properties": {
                "id": {
                    "type": "string",
                    "format": "string"
                },
                "affiliate_id": {
                    "type": "integer",
                    "format": "int"
                },
                "affiliate_key": {
                    "type": "string",
                    "format": "string"
                },
                "cancellation_id": {
                    "type": "string",
                    "format": "string"
                },
                "created_at": {
                    "type": "string",
                    "format": "date"
                },
                "_links": {
                    "$ref": "#/definitions/SwaggerSelfLink"
                }
            },
            "type": "object"
        },
        "SwaggerGetManyCancellationTrackSuccessResponse": {
            "properties": {
                "count": {
                    "type": "integer",
                    "format": "int"
                },
                "total": {
                    "type": "integer",
                    "format": "int"
                },
                "_embedded": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/SwaggerGetCancellationTrackSuccessResponse"
                    }
                },
                "_links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/SwaggerLinks"
                    }
                }
            },
            "type": "object"
        },
...
        "SwaggerError": {
            "properties": {
                "message": {
                    "type": "string",
                    "format": "string"
                },
                "path": {
                    "type": "string",
                    "format": "string"
                }
            },
            "type": "object"
        },
        "SwaggerErrorResponse": {
            "properties": {
                "total": {
                    "type": "integer",
                    "format": "int"
                },
                "_embedded": {
                    "$ref": "#/definitions/SwaggerErrors"
                }
            },
            "type": "object"
        },
        "SwaggerErrors": {
            "properties": {
                "errors": {
                    "$ref": "#/definitions/SwaggerError"
                }
            },
            "type": "object"
        },
        "SwaggerLink": {
            "properties": {
                "href": {
                    "type": "string",
                    "format": "string"
                }
            }
        },
        "SwaggerLinks": {
            "properties": {
                "self": {
                    "$ref": "#/definitions/SwaggerLink"
                },
                "first": {
                    "$ref": "#/definitions/SwaggerLink"
                },
                "prev": {
                    "$ref": "#/definitions/SwaggerLink"
                },
                "next": {
                    "$ref": "#/definitions/SwaggerLink"
                },
                "last": {
                    "$ref": "#/definitions/SwaggerLink"
                }
            }
        },
        "SwaggerSelfLink": {
            "properties": {
                "self": {
                    "type": "string",
                    "$ref": "#/definitions/SwaggerLink"
                }
            }
        }
    }
}